# course-work/implementations/homework/openapi.json
{
    "openapi": "3.0.4",
    "info": {
        "title": "StudySprint API",
        "description": "StudySprint is a distributed study planning system that allows users to manage study sessions and study goals. The API supports CRUD operations, validation and structured data management.",
        "contact": {
            "name": "Dian Yordanov"
        },
        "version": "v1"
    },
    "paths": {
        "/api/study-goals": {
            "get": {
                "tags": [
                    "StudyGoal"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "StudyGoal"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateStudyGoalDto"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateStudyGoalDto"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateStudyGoalDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/study-goals/{id}": {
            "get": {
                "tags": [
                    "StudyGoal"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "tags": [
                    "StudyGoal"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateStudyGoalDto"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateStudyGoalDto"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateStudyGoalDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "StudyGoal"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/study-sessions": {
            "get": {
                "tags": [
                    "StudySession"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "StudySession"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateStudySessionDto"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateStudySessionDto"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateStudySessionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/study-sessions/{id}": {
            "get": {
                "tags": [
                    "StudySession"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "tags": [
                    "StudySession"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateStudySessionDto"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateStudySessionDto"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateStudySessionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "StudySession"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserDto"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserDto"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserDto"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserDto"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CreateStudyGoalDto": {
                "required": [
                    "goalTitle"
                ],
                "type": "object",
                "properties": {
                    "goalTitle": {
                        "maxLength": 100,
                        "minLength": 1,
                        "type": "string"
                    },
                    "targetHours": {
                        "type": "number",
                        "format": "double"
                    },
                    "deadline": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "completed": {
                        "type": "boolean"
                    },
                    "priority": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "userId": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "additionalProperties": false
            },
            "CreateStudySessionDto": {
                "required": [
                    "subject",
                    "title"
                ],
                "type": "object",
                "properties": {
                    "title": {
                        "maxLength": 100,
                        "minLength": 1,
                        "type": "string"
                    },
                    "subject": {
                        "maxLength": 50,
                        "minLength": 1,
                        "type": "string"
                    },
                    "durationMinutes": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "sessionDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "difficulty": {
                        "type": "number",
                        "format": "double"
                    },
                    "userId": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "additionalProperties": false
            },
            "CreateUserDto": {
                "required": [
                    "email",
                    "password",
                    "role",
                    "username"
                ],
                "type": "object",
                "properties": {
                    "username": {
                        "maxLength": 50,
                        "minLength": 1,
                        "type": "string"
                    },
                    "email": {
                        "maxLength": 100,
                        "minLength": 1,
                        "type": "string"
                    },
                    "password": {
                        "maxLength": 150,
                        "minLength": 1,
                        "type": "string"
                    },
                    "role": {
                        "maxLength": 20,
                        "minLength": 1,
                        "type": "string"
                    },
                    "age": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "additionalProperties": false
            },
            "UpdateUserDto": {
                "required": [
                    "email",
                    "password",
                    "role",
                    "username"
                ],
                "type": "object",
                "properties": {
                    "username": {
                        "maxLength": 50,
                        "minLength": 1,
                        "type": "string"
                    },
                    "email": {
                        "maxLength": 100,
                        "minLength": 1,
                        "type": "string"
                    },
                    "password": {
                        "maxLength": 100,
                        "minLength": 1,
                        "type": "string"
                    },
                    "role": {
                        "maxLength": 30,
                        "minLength": 1,
                        "type": "string"
                    },
                    "age": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "additionalProperties": false
            }
        }
    },
    "tags": [
        {
            "name": "StudyGoal"
        },
        {
            "name": "StudySession"
        },
        {
            "name": "User"
        }
    ]
}